resume-transporter/pyproject.toml
[project]
name = "resume-transporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pyyaml>=6.0.2",
    "tortoise-orm[asyncpg]>=0.24.2",
]

[dependency-groups]
dev = [
    "pgadmin4>=9.2",
    "ruff>=0.11.4",
]
deploy = [
    "uvicorn>=0.34.0",
]
